Updated documentation to implement CRAN feedback, which is as follows: Please reduce the length of the title to less than 65 characters.
Please add a web reference for the API/webservices in the form
<https:.....> to the description of the DESCRIPTION file with no space
after 'https:' and angle brackets for auto-linking.

Please add \value to .Rd files regarding exported methods and explain
the functions results in the documentation. Please write about the
structure of the output (class) and also what the output means. (If a
function does not return a value, please document that too, e.g.
\value{No return value, called for side effects} or similar)
Missing Rd-tags:
      ave_daily.Rd: \value
      wildvizApp.Rd: \value

You have examples for unexported functions.
    daily_climate() in:
       daily_climate.Rd
    daily_df() in:
       daily_df.Rd
    daily_stations() in:
       daily_df.Rd
       daily_stations.Rd
  Please either omit these examples or export the functions.